MM15=>MM151 Add Document and PTT
